fix/build_package: workspace cleanup when jenkins ci is not set
When `jenkins_run` is not set workspace is cleaned up either if
`dont_clean` flag is set. This removes any previous file storage
in the workspace which may cause the job to fail.

+++ script/build_package.sh
@@ -1135,7 +1135,7 @@ apply_tf_patch() {
 }
 
 # Clear workspace for a local run
-if not_upon "$jenkins_run"; then
+if not_upon "$jenkins_run" && not_upon "$dont_clean"; then
 	rm -rf "$workspace"
 
 	# Clear residue from previous runs
